スプレー木とロープの Norm 対応。
