+++ Makefile
@@ -25,8 +25,8 @@ SRCS = cmd_cmd_invocation.c cmd_cmd_invocation2.c cmd_exec_command.c		\
 	builtin_cd_path.c builtin_cd_chdir.c builtin_cd_cdpath.c				\
 	builtin_export.c builtin_export2.c builtin_pwd.c builtin_unset.c		\
 	str_utils.c shell_initialization.c cmd_exec_builtin.c rope1.c rope2.c	\
-	rope3.c rope4.c splay1.c splay2.c splay3.c splay4.c splay5.c editor1.c	\
-	editor2.c editor3.c editor4.c editor5.c
+	rope3.c rope4.c splay1.c splay2.c splay3.c splay4.c splay5.c splay6.c	\
+	editor1.c editor2.c editor3.c editor4.c editor5.c
 
 OBJS = $(SRCS:.c=.o)
