+++ splay2.c
@@ -12,9 +12,7 @@ t_splay_tree	*splay_zig_right(t_splay_tree *x, t_splay_tree *p)
 	t_splay_tree	*dest;
 
 	dest = splay_create(
-		x->left, x->value, splay_create(x->right, p->value, p->right));
-	/* splay_release(x); */
-	/* splay_release(p); */
+			x->left, x->value, splay_create(x->right, p->value, p->right));
 	return (dest);
 }
 
@@ -33,13 +31,10 @@ t_splay_tree	*splay_zig_zig_right(
 	t_splay_tree	*dest;
 
 	dest = splay_create(
-		x->left, x->value,
-		splay_create(
-			x->right, p->value,
-			splay_create(p->right, g->value, g->right)));
-	/* splay_release(x); */
-	/* splay_release(p); */
-	/* splay_release(g); */
+			x->left, x->value,
+			splay_create(
+				x->right, p->value,
+				splay_create(p->right, g->value, g->right)));
 	return (dest);
 }
 
@@ -66,11 +61,8 @@ t_splay_tree	*splay_zig_zag_right(
 	c = x->right;
 	d = g->right;
 	dest = splay_create(
-		splay_create(a, p->value, b),
-		x->value,
-		splay_create(c, g->value, d));
-	/* splay_release(x); */
-	/* splay_release(p); */
-	/* splay_release(g); */
+			splay_create(a, p->value, b),
+			x->value,
+			splay_create(c, g->value, d));
 	return (dest);
 }
