+++ splay4.c
@@ -1,5 +1,17 @@
 #include "rope.h"
 
+static t_splay_tree	*splay_1_result(t_splay_path *new_path)
+{
+	t_splay_tree	*result;
+
+	new_path->refcount++;
+	result = splay(new_path);
+	result->refcount++;
+	splay_path_release(new_path);
+	result->refcount--;
+	return (result);
+}
+
 static t_splay_tree	*splay_1(t_splay_path *path)
 {
 	t_splay_tree		*x;
@@ -18,13 +30,20 @@ static t_splay_tree	*splay_1(t_splay_path *path)
 	else
 		new_path = splay_path_create(
 				d2, splay_zig_left(x, p), path->next->next);
-	new_path->refcount++;
+	return (splay_1_result(new_path));
+}
 
+static t_splay_tree	*splay_2_result(t_splay_path *path, t_splay_tree *rotated)
+{
 	t_splay_tree	*result;
+	t_splay_path	*new_path;
 
-	result = splay(new_path);
-	result->refcount++;
+	splay_path_init(
+		&new_path, splay_path_create(
+			path->next->next->dir, rotated, path->next->next->next));
+	splay_init(&result, splay(new_path));
 	splay_path_release(new_path);
+	splay_release(rotated);
 	result->refcount--;
 	return (result);
 }
@@ -53,18 +72,7 @@ static t_splay_tree	*splay_2(t_splay_path *path)
 		else
 			splay_init(&rotated, splay_zig_zag_right(x, p, g));
 	}
-
-	t_splay_tree	*result;
-	t_splay_path	*new_path;
-
-	splay_path_init(
-		&new_path, splay_path_create(
-			path->next->next->dir, rotated, path->next->next->next));
-	splay_init(&result, splay(new_path));
-	splay_path_release(new_path);
-	splay_release(rotated);
-	result->refcount--;
-	return (result);
+	return (splay_2_result(path, rotated));
 }
 
 t_splay_tree	*splay(t_splay_path *path)
@@ -76,58 +84,4 @@ t_splay_tree	*splay(t_splay_path *path)
 		return (splay_1(path));
 	}
 	return (path->node);
-}
-
-t_splay_path	*splay_path_left(t_splay_path *path_orig)
-{
-	t_splay_tree	*current;
-	t_splay_path	*path;
-
-	path = path_orig;
-	path->refcount++;
-	current = path->node;
-	if (current->left)
-	{
-		current = current->left;
-		splay_path_assign(&path, splay_path_create(SPLAY_LEFT, current, path));
-		while (current->right)
-		{
-			current = current->right;
-			splay_path_assign(
-				&path, splay_path_create(SPLAY_RIGHT, current, path));
-		}
-	}
-	else if (path->dir == SPLAY_RIGHT)
-		splay_path_assign(&path, path->next);
-	else
-		splay_path_assign(&path, path->next->next);
-	path->refcount--;
-	return (path);
-}
-
-t_splay_path	*splay_path_right(t_splay_path *path_orig)
-{
-	t_splay_tree	*current;
-	t_splay_path	*path;
-
-	path = path_orig;
-	path->refcount++;
-	current = path->node;
-	if (current->right)
-	{
-		current = current->right;
-		splay_path_assign(&path, splay_path_create(SPLAY_RIGHT, current, path));
-		while (current->left)
-		{
-			current = current->left;
-			splay_path_assign(
-				&path, splay_path_create(SPLAY_LEFT, current, path));
-		}
-	}
-	else if (path->dir == SPLAY_LEFT)
-		splay_path_assign(&path, path->next);
-	else
-		splay_path_assign(&path, path->next->next);
-	path->refcount--;
-	return (path);
 }
